Shooting transferred to server side, changed projectile drawing a little.

+++ src/fingerprint/gameplay/objects/projectiles/Projectile.java
@@ -11,6 +11,7 @@ import fingerprint.gameplay.objects.GameObject;
 import fingerprint.gameplay.objects.events.DeleteEntityEvent;
 import java.util.logging.Level;
 import java.util.logging.Logger;
+import org.newdawn.slick.Color;
 import org.newdawn.slick.Graphics;
 import org.newdawn.slick.Image;
 import org.newdawn.slick.SlickException;
@@ -32,6 +33,8 @@ public class Projectile extends CollidingObject{
     private int damage;
     private String guid;
     private boolean destroyed = false;
+    
+    private byte team;
 
     public Projectile(double angle, double speed, double maxTravelDistance, double initX, double initY) {
         super(initX, initY, new Rectangle((float)initX, (float)initX, 2, 2));
@@ -40,9 +43,16 @@ public class Projectile extends CollidingObject{
         this.maxTravelDistance = maxTravelDistance;
         this.currentTravelDistance = 0;
         this.damage = 0;
-        setCollideToTerrain(true);
+        super.setCollideToTerrain(true);
     }
 
+    public Projectile() {
+        super(0, 0, new Rectangle(0,0,2,2));
+        super.setCollideToTerrain(true);
+        this.currentTravelDistance = 0;
+        this.damage = 0;
+    }
+    
     public void setFriendly(boolean friendly) {
         this.friendly = friendly;
     }
@@ -59,18 +69,21 @@ public class Projectile extends CollidingObject{
         this.damage = damage;
     }
     
+    
 
     @Override
     public void draw(Graphics graphics) {
-        try {
+        //try {
             double[] drawingCoords = getDrawingCoordinates();
-            Image projectileImage = new Image("resources/" + image.getFilename());
+            graphics.setColor(Color.black);
+            graphics.fillRect((float)drawingCoords[0],(float)drawingCoords[1],(float)collideShape.getWidth(),(float)collideShape.getHeight());
+            //Image projectileImage = new Image("resources/" + image.getFilename());
             
-            projectileImage.draw((float)drawingCoords[0], (float)drawingCoords[1]);
+            //projectileImage.draw((float)drawingCoords[0], (float)drawingCoords[1]);
             
-        } catch (SlickException ex) {
-            Logger.getLogger(Projectile.class.getName()).log(Level.SEVERE, null, ex);
-        }
+        //} catch (SlickException ex) {
+        //    Logger.getLogger(Projectile.class.getName()).log(Level.SEVERE, null, ex);
+        //}
     }
 
     @Override
@@ -85,7 +98,7 @@ public class Projectile extends CollidingObject{
 
     @Override
     protected void onTerrainCollision() {
-        if(!destroyed){
+        if(!destroyed && false){
             destroyed = true;
             eventBus.post(new DeleteEntityEvent(guid));
         }
@@ -149,5 +162,14 @@ public class Projectile extends CollidingObject{
     public void setGuid(String guid) {
         this.guid = guid;
     }
+
+    public byte getTeam() {
+        return team;
+    }
+
+    public void setTeam(byte team) {
+        this.team = team;
+    }
+    
     
 }
